+++ backend/internal/core/pipeline_integration_test.go
@@ -20,6 +20,8 @@ import (
 	_ "meguri/plugins/fetcher-chromium"
 	_ "meguri/plugins/fetcher-http"
 	_ "meguri/plugins/filter-exclude-selectors"
+	_ "meguri/plugins/filter-exclude-tags"
+	_ "meguri/plugins/filter-include-tags"
 	_ "meguri/plugins/filter-maincontent"
 	_ "meguri/plugins/filter-selector"
 	_ "meguri/plugins/linkextractor-default"